Move conversion helper functions into separate file
