+++ native/src/main/java/io/ballerina/stdlib/mssql/utils/ConverterUtils.java
@@ -20,24 +20,15 @@ package io.ballerina.stdlib.mssql.utils;
 
 import com.microsoft.sqlserver.jdbc.Geometry;
 import io.ballerina.runtime.api.TypeTags;
-import io.ballerina.runtime.api.types.Field;
-import io.ballerina.runtime.api.types.StructureType;
-import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.utils.TypeUtils;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BDecimal;
-import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.stdlib.mssql.Constants;
-import io.ballerina.stdlib.sql.exception.ApplicationError;
 
 import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
-
-import static io.ballerina.runtime.api.utils.StringUtils.fromString;
 
 /**
  * This class implements the utils methods for the MSSQL Datatypes.
@@ -45,13 +36,13 @@ import static io.ballerina.runtime.api.utils.StringUtils.fromString;
  * @since 0.1.0
  */
 public class ConverterUtils {
-    public static Geometry convertPoint(Object value, Object srid) throws ApplicationError, SQLException {
+    public static Geometry convertPoint(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (TypeUtils.getType(value).getTag() == TypeTags.RECORD_TYPE_TAG) {
-            Map<String, Object> pointValue = getRecordData(value);
-            String pointText = getPointText(pointValue);
+            Map<String, Object> pointValue = ConversionHelperUtils.getRecordData(value);
+            String pointText = ConversionHelperUtils.getPointText(pointValue);
             wkt = String.format("POINT (%s)", pointText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: Point");
@@ -65,12 +56,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertLineString(Object value, Object srid) throws SQLException, ApplicationError {
+    public static Geometry convertLineString(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String lineStringText = getLineStringText(((BArray) value).getValues(), ((BArray) value).size());
+            String lineStringText = ConversionHelperUtils.getLineStringText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("LINESTRING (%s)", lineStringText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: LineString");
@@ -84,12 +76,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertCircularString(Object value, Object srid) throws ApplicationError, SQLException {
+    public static Geometry convertCircularString(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String circularStringText = getCircularStringText(((BArray) value).getValues(), ((BArray) value).size());
+            String circularStringText = ConversionHelperUtils.getCircularStringText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("CIRCULARSTRING (%s)", circularStringText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: Circular String");
@@ -103,12 +96,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertCompoundCurve(Object value, Object srid) throws ApplicationError, SQLException {
+    public static Geometry convertCompoundCurve(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String compoundCurveText = getCompoundCurveText(((BArray) value).getValues(), ((BArray) value).size());
+            String compoundCurveText = ConversionHelperUtils.getCompoundCurveText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("COMPOUNDCURVE (%s)", compoundCurveText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: Compound Curve");
@@ -122,12 +116,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertPolygon(Object value, Object srid) throws ApplicationError, SQLException {
+    public static Geometry convertPolygon(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String polygonText = getPolygonText(((BArray) value).getValues(), ((BArray) value).size());
+            String polygonText = ConversionHelperUtils.getPolygonText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("POLYGON (%s)", polygonText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: Polygon");
@@ -141,12 +136,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertCurvePolygon(Object value, Object srid) throws ApplicationError, SQLException {
+    public static Geometry convertCurvePolygon(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String curvePolygonText = getCurvePolygonText(((BArray) value).getValues(), ((BArray) value).size());
+            String curvePolygonText = ConversionHelperUtils.getCurvePolygonText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("CURVEPOLYGON (%s)", curvePolygonText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: CurvePolygon");
@@ -160,12 +156,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertMultiPoint(Object value, Object srid) throws SQLException, ApplicationError {
+    public static Geometry convertMultiPoint(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String multiPointText = getMultiPointText(((BArray) value).getValues(), ((BArray) value).size());
+            String multiPointText = ConversionHelperUtils.getMultiPointText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("MULTIPOINT (%s)", String.join(", ", multiPointText));
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: LineString");
@@ -179,12 +176,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertMultiLineString(Object value, Object srid) throws SQLException, ApplicationError {
+    public static Geometry convertMultiLineString(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String multiLineStringText = getMultiLineStringText(((BArray) value).getValues(), ((BArray) value).size());
+            String multiLineStringText = ConversionHelperUtils.getMultiLineStringText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("MULTILINESTRING (%s)", multiLineStringText);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: LineString");
@@ -198,12 +196,13 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertMultiPolygon(Object value, Object srid) throws SQLException, ApplicationError {
+    public static Geometry convertMultiPolygon(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
         } else if (value instanceof BArray) {
-            String multiPolygonTest = getMultiPolygonText(((BArray) value).getValues(), ((BArray) value).size());
+            String multiPolygonTest = ConversionHelperUtils.getMultiPolygonText(((BArray) value).getValues(),
+                    ((BArray) value).size());
             wkt = String.format("MULTIPOLYGON (%s)", multiPolygonTest);
         } else {
             throw new SQLException("Unsupported Value: " + value + " for type: MultiPolygon");
@@ -217,7 +216,7 @@ public class ConverterUtils {
         return Geometry.STGeomFromText(wkt, sridInt);
     }
 
-    public static Geometry convertGeometryString(Object value, Object srid) throws SQLException, ApplicationError {
+    public static Geometry convertGeometryString(Object value, Object srid) throws SQLException {
         String wkt;
         if (value instanceof BString) {
             wkt = value.toString();
@@ -237,9 +236,9 @@ public class ConverterUtils {
                 }
 
                 if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.POINT)) {
-                    Map<String, Object> pointValue = getRecordData(
+                    Map<String, Object> pointValue = ConversionHelperUtils.getRecordData(
                             element.get(io.ballerina.stdlib.sql.Constants.TypedValueFields.VALUE));
-                    String pointText = getPointText(pointValue);
+                    String pointText = ConversionHelperUtils.getPointText(pointValue);
                     stringElements[i] = String.format("POINT (%s)", pointText);
                     continue;
                 }
@@ -249,49 +248,49 @@ public class ConverterUtils {
                 int arraySize = elementArray.size();
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.LINESTRING)) {
-                    String lineStringText = getLineStringText(arrayValues, arraySize);
+                    String lineStringText = ConversionHelperUtils.getLineStringText(arrayValues, arraySize);
                     stringElements[i] = String.format("LINESTRING (%s)", lineStringText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.CIRCULARSTRING)) {
-                    String circularStringText = getCircularStringText(arrayValues, arraySize);
+                    String circularStringText = ConversionHelperUtils.getCircularStringText(arrayValues, arraySize);
                     stringElements[i] = String.format("CIRCULARSTRING (%s)", circularStringText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.COMPOUNDCURVE)) {
-                    String circularStringText = getCompoundCurveText(arrayValues, arraySize);
+                    String circularStringText = ConversionHelperUtils.getCompoundCurveText(arrayValues, arraySize);
                     stringElements[i] = String.format("COMPOUNDCURVE (%s)", circularStringText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.POLYGON)) {
-                    String polygonText = getPolygonText(arrayValues, arraySize);
+                    String polygonText = ConversionHelperUtils.getPolygonText(arrayValues, arraySize);
                     stringElements[i] = String.format("POLYGON (%s)", polygonText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.CURVEPOLYGON)) {
-                    String curvePolygonText = getCurvePolygonText(arrayValues, arraySize);
+                    String curvePolygonText = ConversionHelperUtils.getCurvePolygonText(arrayValues, arraySize);
                     stringElements[i] = String.format("CURVEPOLYGON (%s)", curvePolygonText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.MULTIPOINT)) {
-                    String multiPointText = getMultiPointText(arrayValues, arraySize);
+                    String multiPointText = ConversionHelperUtils.getMultiPointText(arrayValues, arraySize);
                     stringElements[i] = String.format("MULTIPOINT (%s)", multiPointText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.MULTILINESTRING)) {
-                    String multiLineStringText = getMultiLineStringText(arrayValues, arraySize);
+                    String multiLineStringText = ConversionHelperUtils.getMultiLineStringText(arrayValues, arraySize);
                     stringElements[i] = String.format("MULTILINESTRING (%s)", multiLineStringText);
                     continue;
                 }
 
                 if (element.getClass().getName().endsWith("$" + Constants.CustomTypeNames.MULTIPOLYGON)) {
-                    String multiPolygonText = getMultiPolygonText(arrayValues, arraySize);
+                    String multiPolygonText = ConversionHelperUtils.getMultiPolygonText(arrayValues, arraySize);
                     stringElements[i] = String.format("MULTIPOLYGON (%s)", multiPolygonText);
                     continue;
                 }
@@ -326,215 +325,4 @@ public class ConverterUtils {
         return money;
     }
 
-    private static Map<String, Object> getRecordData(Object value) throws ApplicationError {
-        Type type = TypeUtils.getType(value);
-        Map<String, Field> structFields = ((StructureType) type).getFields();
-        int fieldCount = structFields.size();
-        Iterator<Field> fieldIterator = structFields.values().iterator();
-        HashMap<String, Object> structData = new HashMap<>();
-        for (int i = 0; i < fieldCount; i++) {
-            Field field = fieldIterator.next();
-            Object bValue = ((BMap) value).get(fromString(field.getFieldName()));
-            int typeTag = field.getFieldType().getTag();
-            switch (typeTag) {
-                case TypeTags.INT_TAG:
-                case TypeTags.FLOAT_TAG:
-                case TypeTags.STRING_TAG:
-                case TypeTags.BOOLEAN_TAG:
-                case TypeTags.DECIMAL_TAG:
-                    structData.put(field.getFieldName(), bValue);
-                    break;
-                case TypeTags.RECORD_TYPE_TAG:
-                    structData.put(field.getFieldName(), getRecordData(bValue));
-                    break;
-                default:
-                    break;
-            }
-        }
-        return structData;
-    }
-
-    protected static String getPointText(Map<String, Object> pointValue) {
-        double x = ((BDecimal) pointValue.get(Constants.Geometric.X)).decimalValue().doubleValue();
-        double y = ((BDecimal) pointValue.get(Constants.Geometric.Y)).decimalValue().doubleValue();
-        return String.format("%f %f", x, y);
-    }
-
-    protected static String getLineStringText(Object[] points, int numPoints) throws ApplicationError {
-        String[] pointStrings = new String[numPoints];
-
-        // Convert array of points into an array of strings
-        for (int i = 0; i < numPoints; i++) {
-            Map<String, Object> pointValue = getRecordData(points[i]);
-            pointStrings[i] = getPointText(pointValue);
-        }
-        // Combine all points into a line string
-        return String.join(", ", pointStrings);
-    }
-
-    protected static String getCircularStringText(Object[] elements, int numElements)
-            throws ApplicationError {
-        String[] pointStrings = new String[numElements];
-        //Convert array of points into an array of strings
-        for (int i = 0; i < numElements; i++) {
-            Map<String, Object> pointValue = getRecordData(elements[i]);
-            pointStrings[i] = getPointText(pointValue);
-        }
-        // Combine all points into a line string
-        return String.join(", ", pointStrings);
-    }
-
-    protected static String getCompoundCurveText(Object[] elements, int numElements) 
-            throws ApplicationError, SQLException {
-        String[] stringElements = new String[numElements];
-        for (int i = 0; i < numElements; i++) {
-            BObject element = (BObject) elements[i];
-
-            if (element instanceof BString) {
-                stringElements[i] = element.toString();
-                continue;
-            }
-
-            BArray elementArray = (BArray) element.get(io.ballerina.stdlib.sql.Constants.TypedValueFields.VALUE);
-
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.LINESTRING)) {
-                String lineStringText = getLineStringText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("(%s)", lineStringText);
-                continue;
-            }
-
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.CIRCULARSTRING)) {
-                String circularStringText = getCircularStringText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("CIRCULARSTRING (%s)", circularStringText);
-                continue;
-            }
-
-            throw new SQLException("Unsupported value: " + element + " for type: CompoundCurve");
-
-        }
-        return String.join(", ", stringElements);
-    }
-
-    protected static String getCurvePolygonText(Object[] elements, int numElements)
-            throws ApplicationError, SQLException {
-        String[] stringElements = new String[numElements];
-        for (int i = 0; i < numElements; i++) {
-            BObject element = (BObject) elements[i];
-
-            if (element instanceof BString) {
-                stringElements[i] = element.toString();
-                continue;
-            }
-
-            BArray elementArray = (BArray) element.get(io.ballerina.stdlib.sql.Constants.TypedValueFields.VALUE);
-
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.LINESTRING)) {
-                String lineStringText = getLineStringText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("(%s)", lineStringText);
-                continue;
-            }
-
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.CIRCULARSTRING)) {
-                String circularStringText = getCircularStringText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("CIRCULARSTRING (%s)", circularStringText);
-                continue;
-            }
-
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.COMPOUNDCURVE)) {
-                String circularStringText = getCompoundCurveText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("COMPOUNDCURVE (%s)", circularStringText);
-                continue;
-            }
-
-            throw new SQLException("Unsupported Value: " + element + " for type: Compound Curve");
-
-        }
-        return String.join(", ", stringElements);
-    }
-
-    protected static String getPolygonText(Object[] elements, int numElements)
-            throws ApplicationError, SQLException {
-        String[] stringElements = new String[numElements];
-        for (int i = 0; i < numElements; i++) {
-            BObject element = (BObject) elements[i];
-
-            if (element instanceof BString) {
-                stringElements[i] = element.toString();
-                continue;
-            }
-
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.LINESTRING)) {
-                BArray elementArray = (BArray) element.get(io.ballerina.stdlib.sql.Constants.TypedValueFields.VALUE);
-                String lineStringText = getLineStringText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("(%s)", lineStringText);
-                continue;
-            }
-
-            throw new SQLException("Unsupported Value: " + element + " for type: Polygon");
-        }
-        return String.join(", ", stringElements);
-    }
-
-    protected static String getMultiPointText(Object[] elements, int numElements)
-            throws ApplicationError {
-        String[] stringElements = new String[numElements];
-
-        // Convert array of points into an array of strings
-        for (int i = 0; i < numElements; i++) {
-            Map<String, Object> pointValue = getRecordData(elements[i]);
-            stringElements[i] = String.format("(%s)", getPointText(pointValue));
-        }
-        // Combine all points into a multi-point
-        return String.join(", ", stringElements);
-    }
-
-    protected static String getMultiLineStringText(Object[] elements, int numElements)
-            throws ApplicationError, SQLException {
-        String[] stringElements = new String[numElements];
-
-        // Convert array of lines into an array of strings
-        for (int i = 0; i < numElements; i++) {
-            BObject element = (BObject) elements[i];
-
-            if (element instanceof BString) {
-                stringElements[i] = element.toString();
-                continue;
-            }
-
-            BArray elementArray = (BArray) element.get(io.ballerina.stdlib.sql.Constants.TypedValueFields.VALUE);
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.LINESTRING)) {
-                String lineStringText = getLineStringText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("(%s)", lineStringText);
-                continue;
-            }
-            throw new SQLException("Unsupported Value: " + element + " for type: Compound Curve");
-        }
-        // Combine all lines into a multi-line
-        return String.join(", ", stringElements);
-    }
-
-    protected static String getMultiPolygonText(Object[] elements, int numElements)
-            throws ApplicationError, SQLException {
-        String[] stringElements = new String[numElements];
-
-        // Convert array of polygons into an array of strings
-        for (int i = 0; i < numElements; i++) {
-            BObject element = (BObject) elements[i];
-
-            if (element instanceof BString) {
-                stringElements[i] = element.toString();
-                continue;
-            }
-
-            BArray elementArray = (BArray) element.get(io.ballerina.stdlib.sql.Constants.TypedValueFields.VALUE);
-            if (element.getClass().getName().contains("$" + Constants.CustomTypeNames.POLYGON)) {
-                String polygonText = getPolygonText(elementArray.getValues(), elementArray.size());
-                stringElements[i] = String.format("(%s)", polygonText);
-                continue;
-            }
-            throw new SQLException("Unsupported Value: " + element + " for type: MultiPolygon");
-        }
-        // Combine all polygons into a multi-polygon
-        return String.join(", ", stringElements);
-    }
 }
